Refactor date expression code to allow changing index date
Previously we evaluated all the expressions during init and discarded
the originals.
